Split function out logically among files
