+++ src/metaswitch/clearwater/etcd_tests/utils/test_config_manager.py
@@ -1,39 +1,6 @@
 from metaswitch.clearwater.config_manager import plugin_base
-from metaswitch.clearwater.cluster_manager.plugin_base import SynchroniserPluginBase
-
+from metaswitch.clearwater.config_manager.etcd_synchronizer import EtcdSynchronizer as CfgEtcdSynchronizer
 from threading import Condition
-from time import sleep
-
-class DummyClusterPlugin(SynchroniserPluginBase):
-    def __init__(self, fn):
-        self.value = None
-        self.fn = fn
-
-    def key(self):
-        return "/dummy/"
-
-    def on_cluster_changing(self, cluster_view):
-        self.value = cluster_view
-
-    def on_joining_cluster(self, cluster_view):
-        self.on_cluster_changing(cluster_view)
-
-    def on_new_cluster_config_ready(self, cluster_view):
-        if self.fn:
-            self.fn()
-            self.fn = None
-        self.on_cluster_changing(cluster_view)
-
-    def on_stable_cluster(self, cluster_view):
-        self.on_cluster_changing(cluster_view)
-
-    def on_leaving_cluster(self, cluster_view):
-        self.on_cluster_changing(cluster_view)
-
-    def wait_for(self, value, timeout=300):
-        for i in range(timeout):
-            if self.value != value:
-                sleep(1)
 
 class DummyCfgPlugin(plugin_base.ConfigPluginBase):
     def __init__(self):
@@ -66,3 +33,22 @@ class DummyCfgPlugin(plugin_base.ConfigPluginBase):
         if self.value != expected:
             self.condvar.wait(timeout)
         self.condvar.release()
+
+
+class TestConfigManager(object):
+    def __init__(self, etcd_ip):
+        self.plugin = DummyCfgPlugin()
+        self.s = CfgEtcdSynchronizer(self.plugin, etcd_ip, "local", None)
+        self.s.start_thread()
+
+    def value(self):
+        return self.plugin.value
+
+    def wait_for(self, *args, **kwargs):
+        self.plugin.wait_for(*args, **kwargs)
+
+    def stop(self):
+        self.s.terminate()
+
+
+
